Removed old code and old testcases. TestSuite Status - Success
