+++ ChessMoves/ChessMoves.Tests/ChessMoves.Tests.csproj
@@ -78,7 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ChessBoardTests.cs" />
-    <Compile Include="ChessPieceTests.cs" />
     <Compile Include="RedesignTests\FakePiece.cs" />
     <Compile Include="RedesignTests\MapperTests.cs" />
     <Compile Include="RedesignTests\PawnTests.cs" />
